chore: move openai-specific item to correct util file
